refactor: extract sub-components from UpdatesTabRequestItem
