+++ web/src/components/feedback/UpdatesTabRequestItem.tsx
@@ -4,7 +4,6 @@ import {
   Check,
   Clock,
   ExternalLink,
-  Eye,
   GitPullRequest,
   Loader2,
   MessageSquare,
@@ -14,15 +13,13 @@ import {
 import { useTranslation } from 'react-i18next'
 import { StatusBadge } from '../ui/StatusBadge'
 import { isTriaged, getStatusDescription, type CloseRequestInput, type FeatureRequest } from '../../hooks/useFeatureRequests'
-import { formatRelativeTime, getStatusInfo, PREVIEW_WARMUP_SECONDS, type PreviewResult } from './FeatureRequestTypes'
+import { formatRelativeTime, getStatusInfo } from './FeatureRequestTypes'
 import { sanitizeUrl } from '@/lib/utils/sanitizeUrl'
-import { isValidPreviewUrl } from '../../lib/utils/isValidPreviewUrl'
-import { MS_PER_SECOND } from '../../lib/constants/time'
 import type { RequestCardState, RequestItemProps, RequestStatusInfo } from './UpdatesTab.types'
 import { getVerifiedFixStorageKey, readVerifiedFixState, writeVerifiedFixState } from './updatesTabStorage'
-
-const REOPEN_COMMENT_ROWS = 3
-const REOPEN_COMMENT_MAX_LENGTH = 1000
+import { RequestStatusBadge } from './RequestStatusBadge'
+import { VerificationActions } from './VerificationActions'
+import { DiffPreview } from './DiffPreview'
 
 export function RequestItem({
   request,
@@ -159,7 +156,7 @@ export function RequestItem({
             />
           )}
           {isAwaitingVerification && (
-            <FixVerificationPrompt
+            <VerificationActions
               requestId={request.id}
               canPerformActions={canPerformActions}
               isLoading={isLoading}
@@ -184,7 +181,7 @@ export function RequestItem({
           )}
 
           {(request.status === 'fix_ready' || request.status === 'fix_complete') && (
-            <PreviewSection
+            <DiffPreview
               request={request}
               previewChecking={previewChecking}
               previewResults={previewResults}
@@ -262,9 +259,7 @@ function UntriagedRequestContent({
         {request.request_type === 'bug' ? '🐛 ' : '✨ '}{request.title}
       </p>
       <div className="flex items-center gap-2 mt-1.5 flex-wrap">
-        <span className={`px-1.5 py-0.5 text-2xs font-medium rounded ${statusInfo.bgColor} ${statusInfo.color}`}>
-          {statusInfo.label}
-        </span>
+        <RequestStatusBadge label={statusInfo.label} className={`${statusInfo.bgColor} ${statusInfo.color}`} />
         {request.github_issue_url && (
           <ExternalLinkRow href={request.github_issue_url} colorClass="text-purple-400 hover:text-purple-300">
             <ExternalLink className="w-3 h-3" />
@@ -278,9 +273,7 @@ function UntriagedRequestContent({
     </>
   ) : (
     <div className="flex items-center gap-2 mt-1.5 flex-wrap">
-      <span className={`px-1.5 py-0.5 text-2xs font-medium rounded ${statusInfo.bgColor} ${statusInfo.color}`}>
-        {statusInfo.label}
-      </span>
+      <RequestStatusBadge label={statusInfo.label} className={`${statusInfo.bgColor} ${statusInfo.color}`} />
       <span className="text-xs text-muted-foreground italic">Awaiting maintainer attention</span>
       {request.github_issue_number && <span className="text-xs text-muted-foreground">#{request.github_issue_number}</span>}
       {request.github_issue_url && (
@@ -312,18 +305,12 @@ function TriagedRequestContent({
         {request.request_type === 'bug' ? '🐛 ' : '✨ '}{request.title}
       </p>
       <div className="flex items-center gap-2 mt-1.5 flex-wrap">
-        <span className={`px-1.5 py-0.5 text-2xs font-medium rounded ${statusInfo.bgColor} ${statusInfo.color}`}>
-          {statusInfo.label}
-        </span>
+        <RequestStatusBadge label={statusInfo.label} className={`${statusInfo.bgColor} ${statusInfo.color}`} />
         {request.status === 'fix_complete' && (
-          <span className="px-1.5 py-0.5 text-2xs font-medium rounded bg-muted text-muted-foreground">
-            Closed
-          </span>
+          <RequestStatusBadge label="Closed" className="bg-muted text-muted-foreground" />
         )}
         {isAwaitingVerification && (
-          <span className="px-1.5 py-0.5 text-2xs font-medium rounded bg-blue-500/20 text-blue-300">
-            {t('feedback.awaitingVerificationBadge')}
-          </span>
+          <RequestStatusBadge label={t('feedback.awaitingVerificationBadge')} className="bg-blue-500/20 text-blue-300" />
         )}
         {getStatusDescription(request.status, request.closed_by_user) && (
           <span className={`text-xs text-muted-foreground ${shouldBlur ? 'blur-xs select-none' : ''}`}>
@@ -354,14 +341,10 @@ function FixCompleteBanner({
           <span className="text-xs font-semibold text-green-400">{t('feedback.fixMerged')}</span>
         </div>
         {isVerified && (
-          <span className="px-1.5 py-0.5 text-2xs font-medium rounded bg-green-500/20 text-green-300">
-            {t('feedback.verifiedByYou')}
-          </span>
+          <RequestStatusBadge label={t('feedback.verifiedByYou')} className="bg-green-500/20 text-green-300" />
         )}
         {isAwaitingVerification && (
-          <span className="px-1.5 py-0.5 text-2xs font-medium rounded bg-blue-500/20 text-blue-300">
-            {t('feedback.awaitingVerificationBadge')}
-          </span>
+          <RequestStatusBadge label={t('feedback.awaitingVerificationBadge')} className="bg-blue-500/20 text-blue-300" />
         )}
       </div>
       <p className="text-xs text-green-300/80 mb-2">
@@ -393,180 +376,6 @@ function FixCompleteBanner({
       </div>
     </div>
   )
-}
-
-function FixVerificationPrompt({
-  requestId,
-  canPerformActions,
-  isLoading,
-  isReopenFormVisible,
-  reopenComment,
-  onVerify,
-  onToggleReopenForm,
-  onReopenCommentChange,
-  onReopenSubmit,
-  onShowLoginPrompt,
-}: {
-  requestId: string
-  canPerformActions: boolean
-  isLoading: boolean
-  isReopenFormVisible: boolean
-  reopenComment: string
-  onVerify: () => void
-  onToggleReopenForm: () => void
-  onReopenCommentChange: (value: string) => void
-  onReopenSubmit: () => void
-  onShowLoginPrompt: () => void
-}) {
-  const { t } = useTranslation()
-  const isCommentEmpty = reopenComment.trim().length === 0
-
-  return (
-    <div className="mt-2 rounded-lg border border-blue-500/30 bg-blue-500/10 p-3" data-testid={`awaiting-verification-${requestId}`}>
-      <p className="text-sm font-medium text-blue-200">{t('feedback.awaitingVerificationQuestion')}</p>
-      <div className="mt-2 flex flex-wrap gap-2">
-        <button
-          onClick={canPerformActions ? onVerify : onShowLoginPrompt}
-          disabled={canPerformActions && isLoading}
-          className="px-2.5 py-1.5 text-xs rounded bg-green-500/20 hover:bg-green-500/30 text-green-300 transition-colors disabled:opacity-50"
-        >
-          {canPerformActions && isLoading ? t('feedback.verifyingFix') : t('feedback.verifyFix')}
-        </button>
-        <button
-          onClick={canPerformActions ? onToggleReopenForm : onShowLoginPrompt}
-          className="px-2.5 py-1.5 text-xs rounded bg-secondary hover:bg-secondary/80 text-foreground transition-colors"
-        >
-          {t('feedback.stillBroken')}
-        </button>
-      </div>
-      {isReopenFormVisible && (
-        <div className="mt-3 space-y-2">
-          <textarea
-            value={reopenComment}
-            onChange={(event) => onReopenCommentChange(event.target.value.slice(0, REOPEN_COMMENT_MAX_LENGTH))}
-            rows={REOPEN_COMMENT_ROWS}
-            maxLength={REOPEN_COMMENT_MAX_LENGTH}
-            className="w-full rounded-md border border-border bg-background/70 px-3 py-2 text-sm text-foreground outline-none focus:border-blue-400"
-            placeholder={t('feedback.stillBrokenPlaceholder')}
-          />
-          <div className="flex flex-wrap items-center gap-2">
-            <button
-              onClick={onReopenSubmit}
-              disabled={isLoading || isCommentEmpty}
-              className="px-2.5 py-1.5 text-xs rounded bg-blue-500/20 hover:bg-blue-500/30 text-blue-200 transition-colors disabled:opacity-50"
-            >
-              {isLoading ? t('feedback.submittingReopen') : t('feedback.submitStillBroken')}
-            </button>
-            <button
-              onClick={onToggleReopenForm}
-              className="px-2.5 py-1.5 text-xs rounded bg-secondary hover:bg-secondary/80 text-muted-foreground transition-colors"
-            >
-              {t('actions.cancel')}
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  )
-}
-
-function PreviewSection({
-  request,
-  previewChecking,
-  previewResults,
-  onCheckPreview,
-}: {
-  request: FeatureRequest
-  previewChecking: number | null
-  previewResults: Record<number, PreviewResult>
-  onCheckPreview: (prNumber: number) => Promise<void>
-}) {
-  const checkedPreview = request.pr_number ? previewResults[request.pr_number] : null
-  const previewUrl = request.netlify_preview_url || (checkedPreview?.status === 'ready' ? checkedPreview.preview_url : null)
-  const safePreviewUrl = isValidPreviewUrl(previewUrl) ? previewUrl : null
-  const isCheckingThis = previewChecking === request.pr_number
-  const readyAt = checkedPreview?.ready_at ? new Date(checkedPreview.ready_at) : null
-  const secondsSinceReady = readyAt ? (Date.now() - readyAt.getTime()) / MS_PER_SECOND : Infinity
-  const isWarmingUp = secondsSinceReady < PREVIEW_WARMUP_SECONDS
-
-  if (safePreviewUrl && request.status === 'fix_ready') {
-    if (isWarmingUp) {
-      const secondsLeft = Math.ceil(PREVIEW_WARMUP_SECONDS - secondsSinceReady)
-      return (
-        <div className="mt-2 p-2 bg-yellow-500/10 border border-yellow-500/30 rounded">
-          <div className="flex items-center gap-2">
-            <Loader2 className="w-4 h-4 text-yellow-400 animate-spin" />
-            <span className="text-xs text-yellow-400 font-medium">Preview warming up... ({secondsLeft}s)</span>
-          </div>
-        </div>
-      )
-    }
-
-    return (
-      <div className="mt-2 p-2 bg-green-500/10 border border-green-500/30 rounded">
-        <div className="flex items-center justify-between gap-2">
-          <div className="flex items-center gap-2">
-            <Eye className="w-4 h-4 text-green-400" />
-            <span className="text-xs text-green-400 font-medium">Preview Available</span>
-          </div>
-          <button
-            onClick={(event) => {
-              event.stopPropagation()
-              if (isValidPreviewUrl(safePreviewUrl)) {
-                window.open(safePreviewUrl, '_blank', 'noopener,noreferrer')
-              }
-            }}
-            className="px-2 py-1 text-xs rounded bg-green-500 hover:bg-green-600 text-white transition-colors flex items-center gap-1"
-          >
-            <ExternalLink className="w-3 h-3" />
-            Try It
-          </button>
-        </div>
-      </div>
-    )
-  }
-
-  if (safePreviewUrl) {
-    return (
-      <button
-        onClick={(event) => {
-          event.stopPropagation()
-          if (isValidPreviewUrl(safePreviewUrl)) {
-            window.open(safePreviewUrl, '_blank', 'noopener,noreferrer')
-          }
-        }}
-        className="text-xs text-green-400 hover:text-green-300 flex items-center gap-1 mt-1 bg-transparent border-0 p-0 cursor-pointer"
-      >
-        <Eye className="w-3 h-3" />
-        Preview
-      </button>
-    )
-  }
-
-  if (request.pr_number && request.status === 'fix_ready') {
-    return (
-      <div className="mt-1.5 flex items-center gap-2">
-        <button
-          onClick={(event) => {
-            event.stopPropagation()
-            void onCheckPreview(request.pr_number!)
-          }}
-          disabled={isCheckingThis}
-          className="text-xs text-muted-foreground hover:text-green-400 flex items-center gap-1 transition-colors disabled:opacity-50"
-        >
-          {isCheckingThis ? <Loader2 className="w-3 h-3 animate-spin" /> : <Eye className="w-3 h-3" />}
-          Check Preview
-        </button>
-        {checkedPreview && checkedPreview.status !== 'ready' && (
-          <span className="text-2xs text-muted-foreground">
-            {checkedPreview.status === 'pending' ? 'Building...' : checkedPreview.message || checkedPreview.status}
-          </span>
-        )}
-      </div>
-    )
-  }
-
-  return null
 }
 
 function RequestActions({
